# docs/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CivicPulse API",
    "description": "AI Development-Needs Intelligence Layer \u2014 Code for Communities 2.0 Track 1. Intake, Multilingual NLP, Clustering, Geospatial, Evidence, Gap Detection, and Policy Simulation. Contract: docs/api/API_CONTRACT.md",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clusters": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "List clusters, filterable by sector/district/status \u2014 FR-071 (analyst+)",
        "description": "List clusters with multi-factor filtering.",
        "operationId": "list_clusters_clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "issue_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Type"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Full cluster detail (analyst+)",
        "operationId": "get_cluster_clusters__cluster_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "clusters"
        ],
        "summary": "Correct classification/location \u2014 FR-060 (reviewer+)",
        "operationId": "correct_cluster_clusters__cluster_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterCorrection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}/review": {
      "post": {
        "tags": [
          "clusters"
        ],
        "summary": "Approve/reject/request-more-evidence \u2014 FR-059 (reviewer+)",
        "operationId": "review_cluster_clusters__cluster_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}/evidence": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Evidence panel payload \u2014 FR-051 (analyst+)",
        "operationId": "get_evidence_clusters__cluster_id__evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePanel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}/gap-analysis": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Gap detection result \u2014 FR-040\u2013044 (analyst+)",
        "operationId": "get_gap_analysis_clusters__cluster_id__gap_analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters/{cluster_id}/priority": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Priority factor breakdown \u2014 FR-050 (analyst+)",
        "operationId": "get_priority_clusters__cluster_id__priority_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriorityBreakdown"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geospatial/clusters": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "GeoJSON for map rendering \u2014 FR-029 (analyst+)",
        "description": "Returns valid GeoJSON points for clusters. Unresolved locations have geometry: None (FR-021/033).",
        "operationId": "cluster_geojson_geospatial_clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClusterCorrection": {
        "properties": {
          "issue_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Type"
          },
          "location_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Notes"
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "note"
        ],
        "title": "ClusterCorrection"
      },
      "ClusterDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "independent_demand_count": {
            "type": "integer",
            "title": "Independent Demand Count"
          },
          "raw_message_count": {
            "type": "integer",
            "title": "Raw Message Count"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "location": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "uncertainty_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainty Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "issue_type",
          "status",
          "independent_demand_count",
          "raw_message_count"
        ],
        "title": "ClusterDetail"
      },
      "ClusterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClusterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ClusterListResponse"
      },
      "ClusterSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "independent_demand_count": {
            "type": "integer",
            "title": "Independent Demand Count"
          },
          "raw_message_count": {
            "type": "integer",
            "title": "Raw Message Count"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          }
        },
        "type": "object",
        "required": [
          "id",
          "issue_type",
          "status",
          "independent_demand_count",
          "raw_message_count"
        ],
        "title": "ClusterSummary"
      },
...
      "EvidencePanel": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "independent_demand_count": {
            "type": "integer",
            "title": "Independent Demand Count"
          },
          "raw_message_count": {
            "type": "integer",
            "title": "Raw Message Count"
          },
          "location": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "infrastructure_context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infrastructure Context"
          },
          "existing_project_check": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Project Check"
          },
          "priority_factors": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Factors"
          },
          "uncertainty_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainty Notes"
          },
          "dataset_versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dataset Versions"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "issue_type",
          "independent_demand_count",
          "raw_message_count"
        ],
        "title": "EvidencePanel",
        "description": "PRD \u00a710.3 shape. Every number traceable to a stored field or formula (FR-052)."
      },
      "GapAnalysisResult": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "gap_found": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Found"
          },
          "demand_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demand Summary"
          },
          "gap_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Summary"
          },
          "recommendation_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Summary"
          },
          "benchmark_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Used"
          },
          "conflicting_project": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflicting Project"
          },
          "uncertainty_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainty Notes"
          }
        },
        "type": "object",
        "required": [
          "cluster_id"
        ],
        "title": "GapAnalysisResult"
      },
...
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "title": "GeoJSONFeatureCollection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriorityBreakdown": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "demand": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demand"
          },
          "gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          },
          "impact": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "equity_adjustment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity Adjustment"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          },
          "priority_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Index"
          },
          "is_incomplete": {
            "type": "boolean",
            "title": "Is Incomplete",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "cluster_id"
        ],
        "title": "PriorityBreakdown"
      },
...
      "ReviewActionCreate": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(approve|reject|request_more_evidence)$",
            "title": "Action"
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "action",
          "note"
        ],
        "title": "ReviewActionCreate"
      },
      "ReviewActionResponse": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "audit_log_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Log Id"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "action",
          "review_status"
        ],
        "title": "ReviewActionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}